Move MainActivity class to 'ui' package
