+++ app/src/androidTest/kotlin/mobi/akesson/jdm/ExampleTest.kt
@@ -2,6 +2,7 @@ package mobi.akesson.jdm
 
 import android.support.test.rule.ActivityTestRule
 import android.support.test.runner.AndroidJUnit4
+import mobi.akesson.jdm.ui.MainActivity
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
